requirements.txt
Python 3
MySQL
Django
djangorestframework
